Updated the implementation of Circuit's API's and added README
In the new version, which is used in Coursera (coursera.org),
the implementation design is bit updated from the stariwaybook.
Those new APIs are marked with `New` e.g., `runNew` instead of
removing the original implementation.

The added README.md explains how to use this package outside of
the repository, which should be useful for Scala learner.

project.sbt was added to allow `sbt` to find files for Eclipse.